Add unless to statement.

+++ manifests/vc/standalone.pp
@@ -125,14 +125,17 @@ class sunet::vc::standalone(
   }
 
   sunet::remote_file { "/tmp/safenetauthenticationclient-core.zip":
-       remote_location => $safenetauthenticationclient_core_url,
-       mode            => "0600"
+      remote_location => $safenetauthenticationclient_core_url,
+      mode            => "0600"
+      unless  => "test -f /tmp/safenetauthenticationclient-core.zip",
     } ->
     exec {"Unpack safenetauthenticationclient-core":
-      command => "unzip  /tmp/safenetauthenticationclient-core.zip",
+        command => "unzip  /tmp/safenetauthenticationclient-core.zip",
+        unless  => "test -f /tmp/safenetauthenticationclient-core.zip",
     } ->
     exec {"Install safenetauthenticationclient-core":
-       command => "apt-get install 'SAC 10.8.28 GA Build/Installation/withoutUI/Ubuntu-2004/safenetauthenticationclient-core_10.8.28_amd64.deb' -y"
+        command => "apt-get install 'SAC 10.8.28 GA Build/Installation/withoutUI/Ubuntu-2004/safenetauthenticationclient-core_10.8.28_amd64.deb' -y"
+        unless  => "test -f /tmp/safenetauthenticationclient-core.zip",
     }
 
   # Compose
